Only show example format for the dynamic badge that is selected
