+++ core/server/error-pages/500.html
@@ -1,6 +1,5 @@
-<!doctype html><meta charset=utf-8><title>A server error occurred</title>
+<!DOCTYPE html><meta charset="utf-8" /><title>A server error occurred</title>
 
-<h1> I have nothing to offer for this request </h1>
-<p> … but blood, toil, tears and sweat.
-</p>
+<h1>I have nothing to offer for this request</h1>
+<p>… but blood, toil, tears and sweat.</p>
 <p>And trying again later.</p>
